# openAPI/ikuzports-openAPI.yml
openapi: 3.0.0
info:
  version: 1.0.0
  title: Ikuzports - API
  description: Api documentation for Ikuzports
servers:
  # Added by API Auto Mocking Plugin
  - description: SwaggerHub API Auto Mocking
    url: https://virtserver.swaggerhub.com/RAMADINAAINIRIZKI/Ikuzports/1.0.0
paths: 
...
  /events:
    post:
      tags: 
       - events
      summary: Insert events data
      security: 
        - JWTAuth: []
      requestBody:
        required: true
        content: 
          multipart/form-data:
            schema:
              type: object
              properties:
                name:
                  type: string
                  example: "Pertandingan Sepak Bola Bandung Raya"
                address:
                  type: string
                  example: "Lapangan Gasibu"
                city:
                  type: string
                  example: "Bandung"
                category_id:
                  type: integer
                  example: 1
                start_date:
                  type: string
                  example: "2023-1-1"
                end_date:
                  type: string
                  example: "2023-1-2"
                image_event:
                  type: string
                  format: binary
                  example: "link"
                maximum_people:
                  type: integer
                  example: 20
                  
            encoding:
              image_event:
                contentType: images/png, images/jpeg
      responses: 
        '200':
          description: Insert success
          content: 
            application/json:
             schema:
              type: object
              example:
                status: "success"
                message: "success insert user data"
        '400':
          $ref: '#/components/responses/400'          
        '500':
          $ref: '#/components/responses/500'
    
    get:
      tags: 
       - events
      summary: Get all event data
      parameters:
        - name: club_category
          description: "category club"
          required: false
          in: query
          schema: 
            type: string
            example: "Sepak Bola"
        - name: city
          description: "city"
          required: false
          in: query
          schema: 
            type: string
            example: "bandung"
      responses:
        '200':
          description: List all event data success
          content:
            application/json:
              schema:
                type: object
                example: 
                  - id: 1
                    name: "Pertandingan Sepak Bola Bandung Raya"
                    address: "Lapangan Gasibu"
                    city: "Bandung"
                    category_id: 1
                    start_date: "2023-1-1"
                    end_date: "2023-1-2"
                    total_participant: 11
                    organizer_name: "Ronaldo"
                    image_event: "link"
                    maximum_people: 20
                  - id: 2
                    name: "Pertandingan Basket Bandung Raya"
                    address: "Lapangan Arcamanik"
                    city: "Bandung"
                    category_id: 2
                    start_date: "2023-1-5"
                    end_date: "2023-1-5"
                    total_participant: 2
                    organizer_name: "Yao Ming"
                    image_event: "link"
                    maximum_people: 25
        '400':
          $ref: '#/components/responses/400'
        '500':
          $ref: '#/components/responses/500'
  /events/{events_id}: 
    get:
      tags: 
       - events
      summary: Get event data by event ID
      security: 
        - JWTAuth: []
      parameters:
        - name: events_id
          description: "id events"
          required: true
          in: path
          schema: 
            type: integer
            example: 1
      responses:
        '200':
          description: List events data by events ID success
          content:
            application/json:
              schema:
                type: object
                example: 
                  id: 1
                  type: string
                  format: binary
                  example: "link"

        '400':
          $ref: '#/components/responses/400'
        '404':
          $ref: '#/components/responses/404' 
        '500':
          $ref: '#/components/responses/500'
...
  /products:
    get:
      tags: 
       - products
      summary: Get all product data
      security: 
        - JWTAuth: []
      parameters:
        - name: product_category
          description: "category product"
          required: false
          in: query
          schema: 
            type: string
            example: "sepatu"
        - name: city
          description: "city"
          required: false
          in: query
          schema: 
            type: string
            example: "bandung"
      responses:
        '200':
          description: List all product data success
          content:
            application/json:
              schema:
                type: object
                example: 
                  - id: 1
                    name: "Sepatu Bola"
                    price: 250000
                    quantity: 6
                    description: "Sepatu bagus"
                    user_id: 1
                    category_id: 1
                    city: "Bandung"
                  - id: 2
                    name: "Jersey"
                    price: 75000
                    quantity: 10
                    description: "Jersey Langka"
                    user_id: 1
                    category_id: 2
                    city: "Cimahi"
        '400':
          $ref: '#/components/responses/400'
        '500':
          $ref: '#/components/responses/500'
    
    post:
      tags: 
       - products
      summary: Insert product data
      security: 
        - JWTAuth: []
      requestBody:
        required: true
        content: 
          application/json:
            schema:
              type: object
              example: 
                name: "Jersey"
                price: 75000
                quantity: 10
                description: "Jersey Langka"
                category_id: 2
                city: "Bandung"
      responses: 
        '200':
          description: Insert success
          content: 
            application/json:
             schema:
              type: object
              example:
                status: "success"
                message: "success insert activities data"
        '400':
          $ref: '#/components/responses/400'          
        '500':
          $ref: '#/components/responses/500'
  /products/{products_id}:
    get:
      tags: 
       - products
      summary: Get all product data
      security: 
        - JWTAuth: []
      parameters:
        - name: products_id
          description: "id products"
          required: true
          in: path
          schema: 
            type: integer
            example: 1
      responses:
        '200':
          description: List product data by galery id success
          content:
            application/json:
              schema:
                type: object
                example: 
                  id: 2
                  name: "Jersey"
                  price: 75000
                  quantity: 10
                  description: "Jersey Langka"
                  user_id: 1
                  category_id: 2
                  city: "Bandung"
        '400':
          $ref: '#/components/responses/400'
        '404':
          $ref: '#/components/responses/404'
        '500':
          $ref: '#/components/responses/500'
    
    put: 
      tags: 
       - products
      summary: Insert product data
      security: 
        - JWTAuth: []
      parameters:
        - name: products_id
          description: "id products"
          required: true
          in: path
          schema: 
            type: integer
            example: 1
      requestBody:
        required: true
        content: 
          application/json:
            schema:
              type: object
              example: 
                name: "Jersey"
                price: 75000
                quantity: 10
                description: "Jersey Langka"
                category_id: 2
                city: "Bandung"
      responses: 
        '200':
          description: Update success
          content: 
            application/json:
             schema:
              type: object
              example:
                status: "success"
                message: "success update product data"
        '400':
          $ref: '#/components/responses/400'          
        '500':
          $ref: '#/components/responses/500'
    
    delete:
      tags: 
       - products
      summary: Delete product data by ID
      security: 
        - JWTAuth: []
      parameters:
        - name: products_id
          description: "id product"
          required: true
          in: path
          schema: 
            type: integer
            example: 1
      responses: 
        '200':
          description: Delete success
          content: 
            application/json:
             schema:
              type: object
              example:
                status: "success"
                message: "success delete product data by galery id"
        '400':
          $ref: '#/components/responses/400'
        '404':
          $ref: '#/components/responses/404' 
        '500':
          $ref: '#/components/responses/500'
...
components:  
  securitySchemes:
    JWTAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: "use Token"
  schemas:
    ResponseWithoutData:
      type: object
      properties:
        status:
          type: string
        message:
          type: string
...
  responses:
      '404':
        description: Not Found
        content:
          application/json:
            schema:
              type: object
              $ref: '#/components/schemas/ResponseWithoutData'
            example: 
              status: "failed"
              message: "data not found"
      '400':
        description: failed.
        content:
          application/json:
            schema:
              type: object
              $ref: '#/components/schemas/ResponseWithoutData'
            example: 
              status: "failed"
              message: "operation failed"
      '500':
        description: Internal Server Error
        content:
          application/json:
            schema:
              type: object
              $ref: '#/components/schemas/ResponseWithoutData'
            example:
              status: "failed"
              message: "internal server error"